third_party/blink/web_tests/external/wpt/editing/event-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Partially-selected editable content: input event
  assert_equals: number of input events fired expected 0 but got 1
[FAIL] Selection spans two editing hosts: input event
  assert_equals: number of input events fired expected 0 but got 1
[FAIL] Changing selection from handler: input event
  assert_equals: event.target expected Element node <div contenteditable="">bar</div> but got Element node <div contenteditable=""><b>foo</b></div>
[FAIL] Command createLink, value "": input event
  assert_equals: number of input events fired expected 1 but got 0
[FAIL] Command fontSize, value "": input event
  assert_equals: number of input events fired expected 1 but got 0
[FAIL] Command fontSize, value "quasit": input event
  assert_equals: number of input events fired expected 1 but got 0
[FAIL] Command formatBlock, value "": input event
  assert_equals: number of input events fired expected 1 but got 0
Harness: the test ran to completion.
